- Added OS X build target. Resolved compilation issues but haven't tested yet.

+++ EX2Kit/Categories/Foundation/NSData+Gzip.m
@@ -25,7 +25,7 @@
     strm.opaque = Z_NULL;
     strm.total_out = 0;
     strm.next_in=(Bytef *)self.bytes;
-    strm.avail_in = self.length;
+    strm.avail_in = (uint)self.length;
     
     // Compresssion Levels:
     //   Z_NO_COMPRESSION
@@ -44,7 +44,7 @@
             [compressed increaseLengthBy: 16384];
         
         strm.next_out = [compressed mutableBytes] + strm.total_out;
-        strm.avail_out = [compressed length] - strm.total_out;
+        strm.avail_out = (uint)([compressed length] - strm.total_out);
         
         deflate(&strm, Z_FINISH);
         
@@ -62,7 +62,7 @@
     if (self.length == 0)
         return nil;
     
-    unsigned full_length = self.length;
+    unsigned full_length = (unsigned)self.length;
     unsigned half_length = self.length / 2;
     
     NSMutableData *decompressed = [NSMutableData dataWithLength: full_length + half_length];
@@ -71,7 +71,7 @@
     
     z_stream strm;
     strm.next_in = (Bytef *)self.bytes;
-    strm.avail_in = self.length;
+    strm.avail_in = (uint)self.length;
     strm.total_out = 0;
     strm.zalloc = Z_NULL;
     strm.zfree = Z_NULL;
@@ -85,7 +85,7 @@
         if (strm.total_out >= [decompressed length])
             [decompressed increaseLengthBy: half_length];
         strm.next_out = [decompressed mutableBytes] + strm.total_out;
-        strm.avail_out = [decompressed length] - strm.total_out;
+        strm.avail_out = (uint)([decompressed length] - strm.total_out);
         
         // Inflate another chunk.
         status = inflate (&strm, Z_SYNC_FLUSH);
